refactor(prompts): migrate acceptance orphan prompt functions into AcceptancePromptBuilder
Four files in src/acceptance/ contained prompt-building logic that violated
the project convention (all LLM prompt assembly must live in src/prompts/builders/).

- acceptance-builder.ts: add buildFixGeneratorPrompt(), buildDiagnosisPrompt(),
  buildRefinementPrompt() + private strategy helpers; export MAX_FILE_LINES
- fix-generator.ts: remove buildFixPrompt() — call builder directly
- refinement.ts: remove buildRefinementPrompt/buildStrategySection/buildRefinedExample
  — call builder directly; remove wrapJsonPrompt import
- fix-diagnosis.ts: remove buildDiagnosisPrompt() wrapper — call builder directly;
  import MAX_FILE_LINES from prompts barrel
- fix-executor.ts: remove buildSourceFixPrompt/buildTestFixPrompt wrappers
  — inline builder calls at the two execute* call sites
- acceptance/index.ts: remove buildRefinementPrompt from public barrel re-export
- prompts/index.ts: export MAX_FILE_LINES, FixGeneratorParams, DiagnosisPromptParams,
  RefinementPromptOptions from acceptance builder
- test files: update imports to use AcceptancePromptBuilder directly
